requirements.txt
easyocr
pytesseract
fitz
pillow
pyzbar
requests
